+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/Suggestions/InstantCommands/GiphyCommandHandler.swift
@@ -38,7 +38,7 @@ public struct GiphyCommandHandler: CommandHandler {
     }
 
     public func canHandleCommand(in text: String, caretLocation: Int) -> ComposerCommand? {
-        if text == id {
+        if text.hasPrefix(id) {
             return ComposerCommand(
                 id: id,
                 typingSuggestion: TypingSuggestion(